+++ RangeHighlight/ModEntry.cs
@@ -12,12 +12,15 @@ namespace RangeHighlight {
         internal ModConfig config;
         internal RangeHighlighter highlighter;
         private IRangeHighlightAPI api;
+        private RangeHighlightAPI _api_private; // for testing non-public stuff
         private DefaultShapes defaultShapes;
+        private IModHelper helper;
 
         public override void Entry(IModHelper helper) {
+            this.helper = helper;
             config = helper.ReadConfig<ModConfig>();
             highlighter = new RangeHighlighter(helper, config);
-            api = new RangeHighlightAPI(this);
+            api = _api_private = new RangeHighlightAPI(this);
             defaultShapes = new DefaultShapes(api);
             installDefaultHighlights();
         }
@@ -38,16 +41,25 @@ namespace RangeHighlight {
             public readonly bool[,] scarecrow;
             public readonly bool[,] deluxeScarecrow;
             public readonly bool[,] junimoHut;
+            public readonly bool[,] cherryBomb;
+            public readonly bool[,] bomb;
+            public readonly bool[,] megaBomb;
 
             public DefaultShapes(IRangeHighlightAPI api) {
                 qualitySprinkler = api.GetSquareCircle(1);
                 iridiumSprinkler = api.GetSquareCircle(2);
                 prismaticSprinkler = api.GetSquareCircle(3);
                 beehouse = api.GetManhattanCircle(5);
-                scarecrow = api.GetCartesianCircle(8);
-                deluxeScarecrow = api.GetCartesianCircle(16);
+                scarecrow = api.GetCartesianCircleWithTruncate(8);
+                deluxeScarecrow = api.GetCartesianCircleWithTruncate(16);
                 junimoHut = api.GetSquareCircle(8);
                 junimoHut[7, 7] = junimoHut[8, 7] = junimoHut[9, 7] = junimoHut[7, 8] = junimoHut[9, 8] = false;
+                cherryBomb = api.GetCartesianCircleWithRound(3, false);
+                bomb = api.GetCartesianCircleWithRound(5, false);
+                megaBomb = api.GetCartesianCircleWithRound(7, false);
+                // yeah, it's strange; but I have the screenshots showing this shape
+                megaBomb[1, 5] = megaBomb[1, 6] = megaBomb[1, 7] = megaBomb[1, 8] = megaBomb[1, 9] = false;
+                megaBomb[13, 5] = megaBomb[13, 6] = megaBomb[13, 7] = megaBomb[13, 8] = megaBomb[13, 9] = false;
             }
 
             public bool[,] GetSprinkler(string name) {
@@ -55,6 +67,12 @@ namespace RangeHighlight {
                 if (name.Contains("quality")) return qualitySprinkler;
                 if (name.Contains("prismatic")) return prismaticSprinkler;
                 return sprinkler;
+            }
+
+            public bool[,] GetBomb(string name) {
+                if (name.Contains("mega")) return megaBomb;
+                if (name.Contains("cherry")) return cherryBomb;
+                return bomb;
             }
         }
 
@@ -88,6 +106,14 @@ namespace RangeHighlight {
                 itemName => {
                     if (itemName.Contains("bee house")) {
                         return new Tuple<Color, bool[,]>(config.BeehouseRangeTint, defaultShapes.beehouse);
+                    } else {
+                        return null;
+                    }
+                });
+            api.AddItemRangeHighlighter("jltaylor-us.RangeHighlight/bomb", null,
+                itemName => {
+                    if (itemName.Contains("bomb")) {
+                        return new Tuple<Color, bool[,]>(config.BombRangeTint, defaultShapes.GetBomb(itemName));
                     } else {
                         return null;
                     }